refactor put game state function in game.h
